simplify api routes for voting, create create-vote and delete-vote POST and DELETE routes
